Update patch set 2
Patch Set 2:

Since the original commit was on Github, I have no idea on how to change the author of the commit to be athurh, and so, I put it in the description. I hope this doesn't cause any inconveniences, and if someone would like to show me how to do it properly, I'd be glad. :-)

Patch-set: 2